Remove git and ssh wrappers.
I keep hitting edge cases where this multifile setup goes wrong for odd reasons
and things crash out with fd errors. There must be a better way to do this, but
no luck with the google thus far. As a stop gap will now be using the
update-configs.sh script to generate a static file periodically.