auto-commit: Allow passing a different server than Google to ping.
It turns out some of the servers where I host auto-committing repositories
don't have a fantastic uptime history, so it's nice to be able to ping them
directly to test whether we are "online" rather than Google. In other words you
want to pass the hostname of the server you are about to try to push to.