ssh-proxy: Suppress yet more errors from processes exiting early.
This makes no sense to me, as we just kicked off 'n' processes so should be
able to wait on 'n' processes. An `OSError` should not get raised here, but
this is indeed showing up in the logs. I can only conclude that I don't have a
fully understanding of the semantics of `wait` or Python's `os.wait` differs
from C's `wait`.