dif: Fix: set CLOEXEC on dsf pipe.
Prior to this commit, we were incorrectly duplicating the file descriptor for
the input pipe to diff-so-fancy when we forked to exec less. As a result,
closing the input pipe in the parent did not HUP diff-so-fancy and it hung
waiting for further input. This commit resolves this by using CLOEXEC as we
should have been doing all along.

Closes Github #24.