rerebase: Strip comments after saving the picklist.
The way Git's interactive rebase is implemented in shell scripting results in
incredibly slow rebases when it is given a picklist with many comment lines.
Unfortunately this meant rerebase often triggered pathological rebases. E.g. one
I had to side step today that ran for 10 minutes without making it through a
single commit.

This commit refactors rerebase to strip comments (both those it has introduced
and the comments Git itself inserted) before handing the picklist back to Git.
The result is our interactive rebases are back up to speed.

Note that my original observation on Github #23 that Git has some kind of no-op
detection logic was actually mistaken. There seems to be no such logic, I just
happened to be working on a fast SSD with a set of commits that applied so fast
it appeared instantaneous.

Closes Github #23.