git-linear: reattach HEAD on `reset`
Previously, the SHA of the initial commit was being stored as the home. The
result was that resetting caused us to checkout a detached head. This was
undesirable if we started from, e.g., "master". We now resolve HEAD if it is a
symbolic reference and then store this as the home pointer. As a result you
should end up reattached when resetting.