kage: fix Clang compilation
Not for the first time, I had to consult Bill Torpey's advice [0] to figure out
what was going wrong. Without the LD flags set correctly, some of the build
tools fail to find libcxx when being executed. I can only assume this worked on
previous machines because a compatible version of libcxx happened to be in the
system directories.

  [0]: http://btorpey.github.io/blog/2015/01/02/building-clang/