replace installation script with a Makefile
Both Bash and GNU Make are not really fit for purpose here, but it turns out
Make sucks marginally less. It would be preferable to use something like Python
(or even Chef/Puppet/Ansible if we were really serious about ops). However, it
seems it is 2018 and I still need access to these tools in BusyBox-like
environments with almost nothing modern available.