rerebase: fix matching picklist lines in commit messages being reworded
Setting rerebase as $GIT_EDITOR has the effect of making it your editor for both
the rebase picklist (intended) and commit message writing when triggered by
git-rebase itself (unintended). This was known all along but assumed benign.
However, what I did not account for was lines in a commit message being reworded
or squashed that looked like picklist entries. Specifically any line that
matched "(pick|fixup|squash) \x+". When seeing this line, rerebase would ask Git
for the diff of the second word, which typically was not an actual Git commit
hash. For example, the commit message that led me to stumble across this was
"squash a -Wsign-compare warning". The effect was the rebase bailing out with a
confusing error message.

The fix here is to correctly tell Git to only use rerebase for the picklist, and
fall back on $EDITOR for writing commit messages.

Github: fixes #43 "rerebase causes strange 'a' argument"