vim: suppress new visual highlighting tweaks
Commit e6d8b4662ddf9356da53f56e363b67b524fd8825¹ of Vim made some changes to
how visual highlighting works that interact poorly with our tweaks. Their
change, when tweaked by us with `cterm=reverse` resulted in near-invisible
highlighting. Ironically, invisible highlighting seems to be one of the problems
the Vim developers are trying to avoid with this and other recent related
changes.

Vim PR #250² has some further commentary relevant to this, including the
puzzling statement:

  We can't rely on fg/bg being different…

I do not understand this because, if the foreground and background colours were
not different, your text would be invisible. In any useful colour scheme, it
seems to me you can rely on the foreground and background colour differing.

This commit does something close to reverting to the old behaviour. Visual
highlighting flips the foreground and background colours, which leads to
something that is both syntax highlighted and readable. Ironically, this is
similar to the screenshot of the 8c default highlighting posted in a PR #250
comment³ that user romainl describes as “an abomination.” I guess beauty is in
the eye of the beholder.

¹ https://github.com/vim/vim/commit/e6d8b4662ddf9356da53f56e363b67b524fd8825
² https://github.com/vim/colorschemes/issues/250
³ https://github.com/vim/colorschemes/issues/250#issuecomment-1848370513