dif: diff-so-fancy-style highlighting
I am not convinced this is worth it. This was surprisingly tricky to implement
and probably still contains a few latent bugs. And the results are
underwhelming. Due to the limited context window and language characteristics
like C-style statements always ending in ';', the word highlighting exhibits
both false positives and false negatives from a user’s intuition. We may want a
way to selectively disable this in future.