dif: more precision in discriminating header
The header lines prefixed with `+++ ` and `--- ` were detected in a bit of an ad
hoc and unreliable way. E.g. an added line that itself started with `+` would
incorrectly be considered a header line. This change makes detection more
accurate, though still not fully precise. To be fully precise we would have to
either track more accurately when we are in the header (`prelude` in the code)
or preserve Git’s original line highlighting.