git: revert to 'diff3' style conflicts
In theory `zdiff3` is supposed to improve your diffs. In practice it is
unavailable in more environments than it is available in. This minor problem
compounds into something worse because `git diff` refuses to run with an
unrecognised style. You cannot even view your edits to this repository while
working on a change like this if Git is using config/.gitconfig and keeps
choking on unrecognised styles.