oi: merge “x = <value of x>” use case into 'oi' macro
The original goal was to have everything accessible through a single `oi` macro
but I was insufficiently creative to figure out how to do it at the time. The
form of this implementation probably explains why. The gymnastic lengths this
new macro goes to are mainly in service of (1) retaining the `printf` format
attribute, without which malformed calls are too easy to make, and (2) avoiding
-Wformat-extra-args compiler warnings that are triggered by most naive
alternatives.