oi: reframe 'oi__' for C++
This essentially completely rethinks the “print? → expr type?” branching logic
for C++. The header is now a bit of a mess – the C and C++ implementations are
almost completely distinct – but compiles under both C and C++ warning-free,
with the exception of ISO C99 warnings for varargs macros with 0 arguments.